# optimize UpSampleNearest 1d 2d and 3d performance on CPU