@mingfeima 2019-12-19
Current `UpSample` implementation has two major performance bottlenecks on CPU:
1. no parallelization
2. given a input of size `NCHW`, the most inner loop is iterating on `NC` which leads to non-contiguous memory access. So even the single core performance is not good.

Benchmark is available from [op_bench-py](https://github.com/mingfeima/op_bench-py). The data is collected on Xeon Skylake 6148, 2*20 cores@2.4GHz.

To reproduce, for inference data: `./run.sh upsample.py`; for training data: `./run.sh upsample --train`

#### single socket inference

operation | input size | original | this pr | speedup
-- | -- | -- | -- | --
upsample_nearest1d | [64, 9, 512] | 3.427 | 0.054 | 63.46
upsample_nearest2d | [8, 9, 64, 64] | 5.880 | 0.103 | 57.09
upsample_nearest3d | [8, 9, 32, 32, 32] | 136.938 | 2.967 | 46.15

#### single core inference

operation | input size | original | this pr | speedup
-- | -- | -- | -- | --
upsample_nearest1d | [64, 9, 512] | 3.056 | 0.516 | 5.92
upsample_nearest2d | [8, 9, 64, 64] | 6.300 | 1.372 | 4.59
upsample_nearest3d | [8, 9, 32, 32, 32] | 145.656 | 43.071 | 3.38

#### dual core training

operation | input size | original | this pr | speedup
-- | -- | -- | -- | --
upsample_nearest1d | [64, 9, 512] | 7.776 | 0.216 | 36.00
upsample_nearest2d | [8, 9, 64, 64] | 13.194 | 0.473 | 27.89
upsample_nearest3d | [8, 9, 32, 32, 32] | 380.244 | 5.963 | 63.77