@mingfeima 2020-02-04
Rebased.

Taking an example of 2d upsample, input of size `NCHW`, the **parallelization pattern** is:

For **forward path**, `cpu_upsample_nearest` parallel on all 4 dimensions `N*C*H*W`;
For **backward path**, `cpu_upsample_nearest_backward` parallel on first two dimensions `N*C`,  since on dimension `H*W`, multiple `grad_output` elements would try to sum onto the same address in `grad_input`, thus dimension `H*W` needs to be sequential here.
The scheme here is similar to mkl-dnn implementation.

Additionally, for commonly used config, fast path is made when computing the nearest index from `output` to `input`:
when input_size == output_size, `input_index = output_index`;
when input_size * 2 == output_size, `input_index = output_index >> 1`.
otherwise go `nearest_neighbor_compute_source_index` which will be slower. I checked with VTune, the int2float and float2int conversion from `nearest_neighbor_compute_source_index` are major hotspots. If we can use a shift, it would be much faster.