@VitalyFedyunin 2020-02-20
Hi! Can you please add benchmark for non-contiguous ( channels last memory layout ) tensors, I want to make sure that we are not degrading performance.