@mingfeima 2020-02-21
@VitalyFedyunin Are you referring a `NHWC` tensor (contiguous) transposed into `NCHW` (non-contiguous)? e.g.
```bash
x = torch.randn(n, c, h, w)
x = x.permute(0, 3, 1, 2)
```
upsample  doesn't have NHWC support yet.