@mingfeima 2020-02-24
> Hi! Can you please add benchmark for non-contiguous ( channels last memory layout ) tensors, I want to make sure that we are not degrading performance.

Add performance numbers for channel last input tensor, this patch is equally effective.
The `nhwc` input tensor is permuted from a `nchw` tensor, `input = input.permute(0, 3, 1, 2)`
```bash
(original)
### using OMP_NUM_THREADS=20
upsample_nearest2d: memory format: nhwc, input size:  torch.Size([8, 9, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration: 6.804 ms

### using OMP_NUM_THREADS=1
upsample_nearest2d: memory format: nhwc, input size:  torch.Size([8, 9, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration: 6.389 ms


(this pr)
### using OMP_NUM_THREADS=20
upsample_nearest2d: memory format: nhwc, input size:  torch.Size([8, 9, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration: 0.209 ms

### using OMP_NUM_THREADS=1
upsample_nearest2d: memory format: nhwc, input size:  torch.Size([8, 9, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration: 2.890 ms
```