@VitalyFedyunin 2020-03-02
Hi @mingfeima can you please fix first suggestion within this PR, as for the second one, it is unnecessary taking into account amazing perf gain we already seeing.