@kimishpatel 2020-03-09
> Hi @mingfeima can you please fix first suggestion within this PR, as for the second one, it is unnecessary taking into account amazing perf gain we already seeing.

@VitalyFedyunin, regarding NHWC format, is my reading of this correct? That the kernel effectively converts the tensor to NCHW contiguous format and output is thus also NCHW contiguous format. We are trying a few models with XNNPACK on mobile where we want NHWC format but it looks like some ops convert them back to NCHW format. max pool was one but plan to use XNNPACK's maxpool impl, and then looking at upsample's impl it seemed that it was also doing format conversion. Ideally we would like to back and forth between NHWC and NCHW. I was doing a simple implementation for NHWC, but wanted to check first if my reading was correct.